Trim all frFont b1d90 calls

+++ src/Battle/btlFormation.c
@@ -837,7 +837,7 @@ extern void func_002c6a00(int param_1,int param_2,u16 param_3);
 extern char func_002c6ba0(int param_1);
 extern undefined * func_002c6e30(int param_1,u32 param_2);
 extern u32 func_002c6f50(u64 param_1,int param_2,u64 param_3,u64 param_4);
-extern void func_002c70d0(u64 param_1,u64 param_2);
+extern void func_002c70d0(int param_1,int param_2);
 extern bool func_002c7250(int param_1);
 extern u32 func_002c7280(int param_1);
 extern u32 func_002c7300(void);
@@ -8822,7 +8822,7 @@ u32 func_002c6f50(u64 param_1,int param_2,u64 param_3,u64 param_4)
 
 // FUN_002c70d0 NONMATCHING
 
-void func_002c70d0(u64 param_1,u64 param_2)
+void func_002c70d0(int param_1,int param_2)
 
 {
   char cVar1 = 0;
